research/table_query/requirements.txt
replicate
azure-ai-formrecognizer
#azure-ai-formrecognizer --pre --upgrade
pillow